backend/requirements.txt
betterpath==0.2.2
blinker==1.9.0
click==8.3.1
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
l==0.11.0
MarkupSafe==3.0.3
pillow==12.1.1
python-dotenv==1.2.2
vcversioner==2.16.0.0
Werkzeug==3.1.6
zope.interface==8.2